refactor: reuse TestRepository class.
